First Pass at Windows Installer
